chore:query db connection pool로 변경 및 추상화
