+++ be/src/infrastructure/query-database/query-db.adapter.ts
@@ -1,5 +1,6 @@
-import { Connection } from 'mysql2/promise';
+import { FieldPacket } from 'mysql2';
+import { RowDataPacket } from 'mysql2/promise';
 
 export interface QueryDBAdapter {
-  connect(identify: string): Promise<Connection>;
+  run(query: string): Promise<{ rows: RowDataPacket[]; fields: FieldPacket[] }>;
 }
